refactor(l5): move source-registry and repo-facts from src/ into server/domain (issue 864)
Two more modules leave src/ for server/domain, each split by responsibility so
every sibling meets the server/ standard whole.

Moved:
- src/source-registry.ts (685) -> server/domain/source-registry.ts (operations)
  + server/domain/source-registry-model.ts (vocabulary, schemas, row mapping,
    namespace/approval helpers, content hashing)
- src/tools/repo-facts.ts (785) -> server/domain/repo-facts.ts (MCP upsert and
  list registration) + server/domain/repo-facts-model.ts (vocabulary, metadata
  schema and published contracts, pure identity/screening helpers)
  + server/domain/repo-facts-source-urls.ts (source-URL trust: the SSRF
    allowlist and the repo/path/commit match)

Both old paths are L5 shims re-exporting the server/ modules, so every src/ and
scripts/ importer keeps its import line untouched.

Server importers repointed: server/tools/source-registry.ts,
server/tools/ingest-conversation-facts.ts, server/tools/repo-facts.ts,
server/contracts/schemas-repo-facts.ts.

Behavior is unchanged. The edits the server/ lint standard forced are all
hoists and table lookups over the same conditions:
- registerSource: the INSERT and the duplicate-identity recovery hoisted out
- updateSource: the SET-clause builder and the zero-row diagnosis hoisted out;
  the eight plain updatable columns now come from one spec table
- isPrivateOrLocalHost: split into the literal-prefix and parsed-octet checks
- sourceUrlMatchesSource: the two host path shapes hoisted into named matchers
- repoFactDedupeKey: the metadata key derivation hoisted out
- both MCP tool handlers hoisted to module scope, with the shared error
  envelope, the upsert authorization, the fact-text screens, the embedding call,
  and the shared-namespace read selection as named functions

insertSourceRow returns the inserted row and refuses rather than fabricating a
record when RETURNING yields none, which replaces the index access the stricter
server/ typing rejects.

src/ runtime closure under server/main.ts: 42 -> 40.

Tests left in src/ (retire or move at L6), all green through the shims:
src/__tests__/source-registry.test.ts,
src/db/migrations/027_source_registry.test.ts,
src/tools/__tests__/source-registry.test.ts,
src/tools/__tests__/repo-facts.test.ts,
src/tools/__tests__/agent-context-pack-repo-facts.test.ts,
src/tools/__tests__/ingest-conversation-facts*.test.ts,
scripts/__tests__/promote-qmd-repo-facts.test.ts.

Done-means: scripts/done-means/864-moved-out-of-src.sh
